requirements.txt
#Django y Django Rest
Django>=2.2.0,<3.0.0
django-filter>=2.2.0,<3.0.0
djangorestframework>=3.10.3,<3.12.0

PyJWT>=1.6.4,<2.1.0
djangorestframework-simplejwt>=4.3.0,<5.0.0

django-cors-headers>=3.2.1,<4.0.0

drf-yasg>=1.13.0,<2.0.0

# psycopg2==2.8.4
psycopg2-binary==2.8.4

# mysqlclient>=2.0.0,<3.00


cryptography>=2.9.2,<3.4

# gunicorn>=20.0.0,<21.0.0
packaging>=19.1,<21.0.0

django-anymail>=7.0.0,<9.0.0

django-environ==0.4.5

factory_boy>=3.0.0,<4.0.0
mock>=4.0.0,<5.0.0
coverage>=5.0.0,<6.0.0


requests>=2.27.0

# celery==5.1.2
redis==3.5.3
pymemcache>=3.5.0

# For Docker Dev
# whitenoise

# For Dev
# ipython
# django-extensions==3.1.0

